Move GlassLabel class to the bundle edu.gemini.shared.gui
